feat: make degit callable like the js version
