+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "degit"
-version = "0.1.2"
+version = "0.1.3"
 authors = ["Vlad Pănăzan <brgdvz@gmail.com>"]
 edition = "2018"
 description = "Rust rewrite of degit"
